+++ packages/validator-ajv8/src/validator.ts
@@ -318,7 +318,7 @@ export default class AJV8Validator<
       return { errors, errorSchema };
     }
 
-    // Include form data with undefined values, which is required for validation.
+    // Include form data with undefined values, which is required for custom validation.
     const newFormData = getDefaultFormState<T>(
       this,
       schema,